converted stories and tests
