+++ components/Form/ObjectInput/ObjectInput.stories.tsx
@@ -1,4 +1,6 @@
-import ObjectInput from './ObjectInput';
+import { Story } from '@storybook/react';
+
+import ObjectInput, { DynamicComponent } from './ObjectInput';
 
 import { useForm } from 'react-hook-form';
 import Button from 'components/Button/Button';
@@ -8,33 +10,20 @@ export default {
   component: ObjectInput,
 };
 
-const Template = (args) => {
+const Template: Story<DynamicComponent> = (args) => {
   const { register, errors, handleSubmit } = useForm();
-  const onSubmit = (data) => {
+  const onSubmit = (data: Record<string, unknown>) => {
     alert(JSON.stringify(data));
   };
 
   return (
     <form onSubmit={handleSubmit(onSubmit)}>
       <ObjectInput
-        name="phone_number"
         label="Phone Number"
-        errors={errors}
-        components={[
-          {
-            component: 'PhoneInput',
-            name: 'phoneNumber',
-            label: 'Phone number',
-            rules: { required: true },
-          },
-          {
-            component: 'TextInput',
-            name: 'phoneType',
-            label: 'Phone type',
-          },
-        ]}
         register={register}
         {...args}
+        name="phone_number"
+        errors={errors}
       />
       <Button className="govuk-button" label="Next" type="submit" />
     </form>
@@ -42,9 +31,25 @@ const Template = (args) => {
 };
 
 export const Default = Template.bind({});
+Default.args = {
+  components: [
+    {
+      component: 'PhoneInput',
+      name: 'phoneNumber',
+      label: 'Phone number',
+      rules: { required: true },
+    },
+    {
+      component: 'TextInput',
+      name: 'phoneType',
+      label: 'Phone type',
+    },
+  ],
+};
 
 export const InLine = Template.bind({});
 InLine.args = {
+  ...Default.args,
   isInline: true,
 };
