+++ components/Form/PhoneInput/PhoneInput.spec.tsx
@@ -1,17 +1,15 @@
 import { fireEvent, render } from '@testing-library/react';
-import NumberInput from './NumberInput';
+import PhoneInput from './PhoneInput';
 
-describe('NumberInput', () => {
-  it('renders a number input', () => {
-    const inputName = 'my-number-input';
+describe('PhoneInput', () => {
+  it('renders a phone input', () => {
+    const inputName = 'my-phone-input';
     const inputLabel = 'My Input';
     const { getByLabelText } = render(
-      <NumberInput name={inputName} label={inputLabel} />
+      <PhoneInput name={inputName} label={inputLabel} />
     );
-
     const labelRegex = new RegExp(`s*${inputLabel}s*`);
-    const input = getByLabelText(labelRegex);
-
+    const input = getByLabelText(labelRegex) as HTMLInputElement;
     expect(input).toBeInTheDocument();
     expect(input.id).toEqual(inputName);
     expect(input.name).toEqual(inputName);
@@ -21,13 +19,11 @@ describe('NumberInput', () => {
     let newValue = '';
     const myAction = jest.fn((e) => (newValue = e.target.value));
     const { getByLabelText } = render(
-      <NumberInput name={'my-input'} label={'My Input'} onChange={myAction} />
+      <PhoneInput name={'my-input'} label={'My Input'} onChange={myAction} />
     );
-
     fireEvent.change(getByLabelText(/\s*My Input\s*/), {
-      target: { value: '2423242526' },
+      target: { value: '07849436768' },
     });
-
-    expect(newValue).toEqual('2423242526');
+    expect(newValue).toEqual('07849436768');
   });
 });
