+++ components/Form/TextInput/TextInput.spec.tsx
@@ -1,16 +1,16 @@
 import { fireEvent, render } from '@testing-library/react';
-import EmailInput from './EmailInput';
+import TextInput from './TextInput';
 
-describe('EmailInput', () => {
-  it('renders a email input', () => {
-    const inputName = 'my-email-input';
+describe('TextInput', () => {
+  it('renders a text input', () => {
+    const inputName = 'my-text-input';
     const inputLabel = 'My Input';
     const { getByLabelText } = render(
-      <EmailInput name={inputName} label={inputLabel} />
+      <TextInput name={inputName} label={inputLabel} />
     );
 
     const labelRegex = new RegExp(`s*${inputLabel}s*`);
-    const input = getByLabelText(labelRegex);
+    const input = getByLabelText(labelRegex) as HTMLInputElement;
 
     expect(input).toBeInTheDocument();
     expect(input.id).toEqual(inputName);
@@ -21,13 +21,13 @@ describe('EmailInput', () => {
     let newValue = '';
     const myAction = jest.fn((e) => (newValue = e.target.value));
     const { getByLabelText } = render(
-      <EmailInput name={'my-input'} label={'My Input'} onChange={myAction} />
+      <TextInput name={'my-input'} label={'My Input'} onChange={myAction} />
     );
 
     fireEvent.change(getByLabelText(/\s*My Input\s*/), {
-      target: { value: 'test@test.com' },
+      target: { value: 'hello' },
     });
 
-    expect(newValue).toEqual('test@test.com');
+    expect(newValue).toEqual('hello');
   });
 });
